separated utils in different classes
